requirements.txt
streamlit>=1.30.0
google-cloud-dialogflow>=2.25.0
google-auth>=2.25.0
protobuf<5.0.0

flask>=2.3.0
gunicorn>=21.2.0